Geometry System

+++ example/EmbeddedBoundary/EmbeddedBoundary.cpp
@@ -33,9 +33,9 @@ int main(int argc, char **argv) {
         argc, argv, [&](std::string const &opt, std::string const &value) -> int {
             if (false) {
             } else if (opt == "n") {
-                num_of_step = static_cast<size_type>(std::atoi(value.c_str()));
+                num_of_step = static_cast<size_type>(std::stol(value.c_str()));
             } else if (opt == "checkpoint") {
-                checkpoint_interval = static_cast<size_type>(std::atoi(value.c_str()));
+                checkpoint_interval = static_cast<size_type>(std::stol(value.c_str()));
             }
             return CONTINUE;
         });
